requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
supervisor==4.2.5
requests==2.31.0
authlib==1.2.1
flask-session==0.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0